Trading works but predictions are not workable.

+++ src/main/java/com/syngleton/chartomancy/analytics/ComputationData.java
@@ -7,8 +7,8 @@ public record ComputationData(
     LocalDateTime endTime,
     ComputationType computationType,
     long computations,
-    int startPricePrediction,
-    int endPricePrediction
+    float startPricePrediction,
+    float endPricePrediction
 ) {
 
 }
